standardize index types

+++ cpp_tests/source/test_cell_lists.cpp
@@ -782,22 +782,22 @@ TEST_F(LatticeNeighborsTest, LargeRcut_Works) {
   std::set<size_t> s(neibs.begin(), neibs.end());
   ASSERT_EQ(neibs.size(), s.size());
 
-  std::vector<std::vector<std::array<std::vector<long> *, 2>>> pairs_inner(
+  std::vector<std::vector<std::array<std::vector<size_t> *, 2>>> pairs_inner(
       lattice.m_nsubdoms);
-  std::vector<std::vector<std::array<std::vector<long> *, 2>>> pairs_boundary(
+  std::vector<std::vector<std::array<std::vector<size_t> *, 2>>> pairs_boundary(
       lattice.m_nsubdoms);
-  std::vector<std::vector<std::vector<long>>> cells(lattice.m_nsubdoms);
+  std::vector<std::vector<std::vector<size_t>>> cells(lattice.m_nsubdoms);
   for (size_t isubdom = 0; isubdom < lattice.m_nsubdoms; isubdom++) {
-    cells[isubdom] = std::vector<std::vector<long>>(
+    cells[isubdom] = std::vector<std::vector<size_t>>(
         lattice.cell_vec_to_global_ind(ncells_vec) / lattice.m_nsubdoms);
   }
   size_t total_cells = 0;
   for (size_t subdom_ncell : lattice.m_subdom_ncells) {
     total_cells += subdom_ncell;
   }
-  std::vector<std::vector<std::vector<long> *>> cell_neighbors(total_cells);
-  for (std::vector<std::vector<long> *> neighbors : cell_neighbors) {
-    neighbors = std::vector<std::vector<long> *>();
+  std::vector<std::vector<std::vector<size_t> *>> cell_neighbors(total_cells);
+  for (std::vector<std::vector<size_t> *> neighbors : cell_neighbors) {
+    neighbors = std::vector<std::vector<size_t> *>();
   }
   lattice.find_neighbor_pairs(pairs_inner, pairs_boundary, cell_neighbors,
                               cells);
@@ -834,22 +834,22 @@ TEST_F(LatticeNeighborsTest, SmallRcut_Works2) {
   std::set<size_t> s(neibs.begin(), neibs.end());
   ASSERT_EQ(neibs.size(), s.size());
 
-  std::vector<std::vector<std::array<std::vector<long> *, 2>>> pairs_inner(
+  std::vector<std::vector<std::array<std::vector<size_t> *, 2>>> pairs_inner(
       lattice.m_nsubdoms);
-  std::vector<std::vector<std::array<std::vector<long> *, 2>>> pairs_boundary(
+  std::vector<std::vector<std::array<std::vector<size_t> *, 2>>> pairs_boundary(
       lattice.m_nsubdoms);
-  std::vector<std::vector<std::vector<long>>> cells(lattice.m_nsubdoms);
+  std::vector<std::vector<std::vector<size_t>>> cells(lattice.m_nsubdoms);
   for (size_t isubdom = 0; isubdom < lattice.m_nsubdoms; isubdom++) {
-    cells[isubdom] = std::vector<std::vector<long>>(
+    cells[isubdom] = std::vector<std::vector<size_t>>(
         lattice.cell_vec_to_global_ind(ncells_vec) / lattice.m_nsubdoms);
   }
   size_t total_cells = 0;
   for (size_t subdom_ncell : lattice.m_subdom_ncells) {
     total_cells += subdom_ncell;
   }
-  std::vector<std::vector<std::vector<long> *>> cell_neighbors(total_cells);
-  for (std::vector<std::vector<long> *> neighbors : cell_neighbors) {
-    neighbors = std::vector<std::vector<long> *>();
+  std::vector<std::vector<std::vector<size_t> *>> cell_neighbors(total_cells);
+  for (std::vector<std::vector<size_t> *> neighbors : cell_neighbors) {
+    neighbors = std::vector<std::vector<size_t> *>();
   }
   lattice.find_neighbor_pairs(pairs_inner, pairs_boundary, cell_neighbors,
                               cells);
@@ -1242,7 +1242,7 @@ public:
 };
 
 template <typename distance_policy>
-size_t get_neighbors(Array<double> &coords, std::vector<long> &iatoms,
+size_t get_neighbors(Array<double> &coords, std::vector<size_t> &iatoms,
                      std::vector<double> &old_coords,
                      pele::CellLists<distance_policy> &cl) {
   stupid_counter<distance_policy::_ndim> counter;
@@ -1255,9 +1255,9 @@ size_t get_neighbors(Array<double> &coords, std::vector<long> &iatoms,
 TEST_F(CellListsSpecificTest, Number_of_neighbors) {
   pele::CellLists<pele::periodic_distance<2>> cell(
       std::make_shared<pele::periodic_distance<2>>(boxvec), boxvec, rcut);
-  std::vector<long> iatoms(0);
+  std::vector<size_t> iatoms(0);
   std::vector<double> old_coords(0);
-  for (long i = 0; i < nparticles; ++i) {
+  for (size_t i = 0; i < nparticles; ++i) {
     iatoms.push_back(i);
     for (size_t idim = 0; idim < ndim; ++idim) {
       old_coords.push_back(x[i * ndim + idim]);
@@ -1275,7 +1275,7 @@ TEST_F(CellListsSpecificTest, Number_of_neighbors) {
 
 template <typename distance_policy>
 size_t
-get_overlaps(Array<double> const &coords, std::vector<long> const &iatoms,
+get_overlaps(Array<double> const &coords, std::vector<size_t> const &iatoms,
              std::vector<double> const &old_coords, Array<double> const &radii,
              pele::CellLists<distance_policy> &cl,
              std::shared_ptr<distance_policy> const &dist) {
@@ -1289,9 +1289,9 @@ get_overlaps(Array<double> const &coords, std::vector<long> const &iatoms,
 TEST_F(CellListsSpecificTest, Number_of_overlaps) {
   auto dist = std::make_shared<pele::periodic_distance<2>>(boxvec);
   pele::CellLists<pele::periodic_distance<2>> cell(dist, boxvec, rcut);
-  std::vector<long> iatoms(0);
+  std::vector<size_t> iatoms(0);
   std::vector<double> old_coords(0);
-  for (long i = 10; i < 40; ++i) {
+  for (size_t i = 10; i < 40; ++i) {
     iatoms.push_back(i);
     for (size_t idim = 0; idim < ndim; ++idim) {
       old_coords.push_back(x[i * ndim + idim]);
